% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_CoxpresDbArchive.R
\name{CoxpresDbAccessor-class}
\alias{CoxpresDbAccessor-class}
\title{Template for \code{CoxpresDbAccessor} class}
\description{
`CoxpresDbAccessor` is a virtual class, to construct an object of this
class, use `CoxpresDbArchiveAccessor` or `CoxpresDbDataframeAccessor`.
}
