% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_CoxpresDbAccessor.R
\name{CoxpresDbArchiveAccessor-class}
\alias{CoxpresDbArchiveAccessor-class}
\title{Constructor for the concrete class `CoxpresDbArchiveAccessor`}
\arguments{
\item{archive}{The path to a CoxpresDb archive (either .tar.bz2
or .tar).}

\item{archive_uncompressed}{The path to an uncompressed copy of the
CoxpresDb archive.}

\item{file_paths}{The relative paths of all file present in
subdirectories of the archive. As a data-frame of `gene_id` -> `file_path`
pairs.}
}
\description{
This is a datastructure that stores the location of a CoxpresDb archive and
the relative paths of all subfiles within that archive. If the archive is
compressed, on construction of \code{CoxpresDbAccessor} a temporary copy of
the uncompressed archive is constructed. So construction may take a couple
of minutes depending on the size of the archive. You may want to wrap a call
to this function in a \code{future()} block.
}
