% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_CoxpresDbArchive.R
\name{CoxpresDbDataframeAccessor-class}
\alias{CoxpresDbDataframeAccessor-class}
\title{Constructor for the concrete class `CoxpresDbDataframeAccessor`}
\arguments{
\item{df}{A dataframe.}
}
\description{
Constructor for the concrete class `CoxpresDbDataframeAccessor`
}
