% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\name{CoxpresDbImporter}
\alias{CoxpresDbImporter}
\title{Constructor for the CoxpresDbImporter class - Use this to obtain data from
a CoxpresDB.jp archive}
\usage{
CoxpresDbImporter(db_archive, temp_dir = tempdir(),
  overwrite_in_bunzip2 = FALSE, remove_in_bunzip2 = FALSE)
}
\arguments{
\item{db_archive}{A file path. Should be a single \code{*.tar.bz2}
or \code{*.tar} file as downloaded from the coxpresdb.jp website.}

\item{temp_dir}{A directory into which a compressed CoxpresDB
archive will be decompressed. By default this is the temp_dir for the
current R session. Only relevant if the coxpresdb archive is compressed.}

\item{overwrite_in_bunzip2}{Boolean. If the CoxpresDB archive is
compressed, and a decompressed copy of the archive is found in the target
directory, should the function throw an exception? See \code{overwrite} in
\code{R.utils::bunzip2}}

\item{remove_in_bunzip2}{Boolean. If a compressed CoxpresDB archive
is provided, should the compressed version be deleted after decompression?
See \code{remove} in \code{R.utils::bunzip2}}
}
\value{
A CoxpresDbImporter object.
}
\description{
The db_archive should be a .tar or a .tar.bz2, if it's a .tar.bz2 then
\code{temp_dir} should be defined and a random copy of the uncompressed
archive will be made. All access to the stored data will be made via the
uncompressed copy of the archive, so make a CoxpresDbImporter _once_ during
any given script.
}
