% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_classes.R
\name{CoxpresDbPartners-class}
\alias{CoxpresDbPartners-class}
\title{Template for CoxpresDbPartners class}
\arguments{
\item{gene_statistics}{ABC}

\item{partners}{DEF}

\item{partner_summaries}{GHI}

\item{cluster_graph}{JKL}
}
\description{
This class stores the gene-partners of a set of genes, and
some summary statistics over those partners that are obtained from analysis
of a user-provided \code{gene_statistics} object.
}
