% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_parse.R
\name{.filter_coex_partners}
\alias{.filter_coex_partners}
\title{Function to filter the contents of a 'coxpresdb.jp' gene-coexpression
dataset}
\usage{
.filter_coex_partners(coex_df, gene_universe, n_partners, mr_threshold)
}
\arguments{
\item{coex_df}{A dataframe containing coexpression data. As
returned by \code{get_all_coex_partners}.}

\item{gene_universe}{The genes in the dataframe should be filtered
to ensure they are all present in this set. Note that both the entries in
\code{source_id} and \code{target_id} are filtered to be in this set.}

\item{n_partners}{The maximum number of partners to return for the
source gene(s) in \code{coex_df}. Gene partners will be sorted in order of
mutual-rank before selecting the top partners.}

\item{mr_threshold}{All gene-pairs in the returned dataset will
have a mutual-rank of at most this value.}
}
\description{
Function to filter the contents of a 'coxpresdb.jp' gene-coexpression
dataset
}
