% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\name{.is_coxpresdb_archive}
\alias{.is_coxpresdb_archive}
\title{Checks whether a provided file-path points to a valid CoxpresDB.jp archive}
\usage{
.is_coxpresdb_archive(db_archive)
}
\arguments{
\item{db_archive}{A file path. Should be a single \code{*.tar.bz2}
or \code{*.tar} file as downloaded from the coxpresdb.jp website.}
}
\description{
Checks whether a provided file-path points to a valid CoxpresDB.jp archive
}
