% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_data_validity.R
\name{.is_gene_statistics_df}
\alias{.is_gene_statistics_df}
\title{Checks that a User-provided gene_statistics data-frame conforms to the
required standard}
\usage{
.is_gene_statistics_df(x)
}
\arguments{
\item{x}{A putative \code{gene_statistics} data-frame}
}
\description{
The data-frame should have columns `gene_id:char`, `p_value:numeric` and
`direction:numeric` and should have at most one row for each gene_id.
}
