% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\docType{methods}
\name{get_all_coex_partners,character,CoxpresDbArchiveAccessor-method}
\alias{get_all_coex_partners,character,CoxpresDbArchiveAccessor-method}
\title{Import all the coexpression partner data for a single gene from a given
CoxpresDb archive.}
\usage{

  \S4method{get_all_coex_partners}{character,CoxpresDbArchiveAccessor}(gene_id,
  importer)
}
\arguments{
\item{gene_id}{A gene-identifier in the same format as present
throughout the CoxpresDb archive. If this isn't a single identifier, or
it isn't present in the database, the function will throw an error.}

\item{importer}{\code{CoxpresDbArchiveAccessor} object.}
}
\value{
A single dataframe containing the source -> target mappings,
  and the mutual ranks between the gene pairs.
}
\description{
Users should use \code{get_coex_partners(gene_ids, importer)} rather than
this (unexported) method.
}
