% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_parse.R
\name{get_coex_partners}
\alias{get_coex_partners}
\title{Returns the coexpression partners of a given set of genes}
\usage{
get_coex_partners(gene_ids, importer, gene_universe = NULL,
  n_partners = 100, mr_threshold = NULL)
}
\arguments{
\item{gene_ids}{A vector of gene identifiers. All of these
should be annotated within the 'coxpresDb' dataset.}

\item{importer}{A \code{CoxpresDbAccessor}. This allows access
to a 'coxpresDb' archive.}

\item{gene_universe}{The genes in the dataframe should be filtered
to ensure they are all present in this set. Note that both the entries in
\code{source_id} and \code{target_id} are filtered to be in this set.}

\item{n_partners}{The maximum number of partners to return for the
source gene(s) in \code{coex_df}. Gene partners will be sorted in order of
mutual-rank before selecting the top partners.}

\item{mr_threshold}{All gene-pairs in the returned dataset will
have a mutual-rank of at most this value.}
}
\description{
Several filters are applied while identifying the coexpression partners of a
given set of genes. Only those gene-pairs with a mutual rank below- the
user-selected threshold, and for a given input gene _at most_
\code{n_partners} are returned.
}
\details{
If the user provides a \code{gene_universe}, the dataframe will only contain
genes that are present in this set (the default universe is the entirety of
the genes present in the coxpresDB archive).

If the thresholds, number-of-partners or gene_universe are set to
\code{NULL}, the corresponding filters are not applied to the dataset.
}
