% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Obtains all the file-paths that are present in a CoxpresDb archive}
\usage{
get_file_paths(x)
}
\arguments{
\item{x}{A \code{CoxpresDbImporter} object corresponding
to a CoxpresDb.jp archive}
}
\value{
A tibble of gene_id:file_path pairs, each file is present in
  the archive (referenced within \code{CoxpresDbImporter}) and represents
  the coexpression data for a particular gene.
}
\description{
Obtains all the file-paths that are present in a CoxpresDb archive
}
