% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\docType{methods}
\name{get_gene_ids,CoxpresDbImporter-method}
\alias{get_gene_ids,CoxpresDbImporter-method}
\title{Imports the gene identifiers that are represented within a given
CoxpresDB.jp archive}
\usage{
\S4method{get_gene_ids}{CoxpresDbImporter}(x)
}
\arguments{
\item{x}{A \code{CoexpresDbImporter} corresponding to a
CoxpresDB archive.}
}
\value{
A vector of gene-ids; data for each such gene is present in
the user-supplied archive
}
\description{
Imports the gene identifiers that are represented within a given
CoxpresDB.jp archive
}
