% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\name{get_gene_ids}
\alias{get_gene_ids}
\title{Generic method for getting gene identifiers.}
\usage{
get_gene_ids(x)
}
\arguments{
\item{x}{A \code{CoexpresDbAccessor} corresponding to a
CoxpresDB archive.}
}
\description{
The returned vector contains all gene IDs present in the dataset: this
should include all source- and target- genes.
}
