% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_classes.R
\docType{methods}
\name{get_gene_universe,CoxpresDbPartners-method}
\alias{get_gene_universe,CoxpresDbPartners-method}
\title{Obtains the set of all genes considered in the analysis of gene-partners}
\usage{
\S4method{get_gene_universe}{CoxpresDbPartners}(x)
}
\arguments{
\item{x}{A CoxpresDbPartners object, as returned by
\code{run_coex_partner_workflow}}
}
\description{
Obtains the set of all genes considered in the analysis of gene-partners
}
