% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_classes.R
\name{get_gene_universe}
\alias{get_gene_universe}
\title{Generic function for returning the set of all genes considered in the
analysis of gene-partners}
\usage{
get_gene_universe(x)
}
\arguments{
\item{x}{A CoxpresDbPartners object, as returned by
\code{run_coex_partner_workflow}}
}
\description{
Generic function for returning the set of all genes considered in the
analysis of gene-partners
}
