% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_classes.R
\docType{methods}
\name{get_source_genes,CoxpresDbPartners-method}
\alias{get_source_genes,CoxpresDbPartners-method}
\title{Obtains the set of source genes for which partner genes were assessed}
\usage{
\S4method{get_source_genes}{CoxpresDbPartners}(x)
}
\arguments{
\item{x}{A CoxpresDbPartners object, as returned by
\code{run_coex_partner_workflow}}
}
\description{
Obtains the set of source genes for which partner genes were assessed
}
