% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_classes.R
\name{get_source_genes}
\alias{get_source_genes}
\title{Generic function for obtaining the set of source genes for which partner
genes were assessed}
\usage{
get_source_genes(x)
}
\arguments{
\item{x}{A CoxpresDbPartners object, as returned by
\code{run_coex_partner_workflow}}
}
\description{
Generic function for obtaining the set of source genes for which partner
genes were assessed
}
