% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\docType{methods}
\name{get_source_ids,CoxpresDbArchiveAccessor-method}
\alias{get_source_ids,CoxpresDbArchiveAccessor-method}
\title{Imports the source-gene identifiers that are represented within a given
CoxpresDB.jp archive}
\usage{
\S4method{get_source_ids}{CoxpresDbArchiveAccessor}(x)
}
\arguments{
\item{x}{A \code{CoexpresDbArchiveAccessor} corresponding
to a CoxpresDB archive.}
}
\value{
A vector of gene-ids; coexpression partners for each such gene
  are available in the user-supplied archive.
}
\description{
Imports the source-gene identifiers that are represented within a given
CoxpresDB.jp archive
}
