% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\docType{methods}
\name{get_source_ids,CoxpresDbDataframeAccessor-method}
\alias{get_source_ids,CoxpresDbDataframeAccessor-method}
\title{Imports the source gene identifiers that are represented within a
data-frame-based CoxpresDB.jp archive}
\usage{
\S4method{get_source_ids}{CoxpresDbDataframeAccessor}(x)
}
\arguments{
\item{x}{A \code{CoxpresDbDataframeAccessor}
corresponding to a CoxpresDB archive.}
}
\value{
A vector of gene-ids; coexpression partners for each such gene
  are present in the user-supplied archive.
}
\description{
Imports the source gene identifiers that are represented within a
data-frame-based CoxpresDB.jp archive
}
