% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\name{get_source_ids}
\alias{get_source_ids}
\title{Generic method for getting source-gene identifiers.}
\usage{
get_source_ids(x)
}
\arguments{
\item{x}{A \code{CoexpresDbAccessor} corresponding to a
CoxpresDB archive.}
}
\description{
The returned vector contains all gene IDs for which coexpression partners
can be obtained from the dataset. There may be target-genes that are not
source-genes.
}
