% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_io.R
\docType{methods}
\name{import_all_coex_partners,character,CoxpresDbAccessor-method}
\alias{import_all_coex_partners,character,CoxpresDbAccessor-method}
\title{Import all the coexpression partner data for a single gene from a given
CoxpresDb archive.}
\usage{
\S4method{import_all_coex_partners}{character,CoxpresDbAccessor}(gene_id,
  importer)
}
\arguments{
\item{gene_id}{A gene-identifier in the same format as present
throughout the CoxpresDb archive.}

\item{importer}{\code{CoxpresDbAccessor} object.}
}
\value{
A single dataframe containing the source -> target mappings,
  and both the mutual ranks and correlations between the gene pairs.
}
\description{
Users should use \code{get_coex_partners(gene_ids, importer)} rather than
this method.
}
