% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_workflows.R
\name{run_coex_partner_workflow}
\alias{run_coex_partner_workflow}
\title{Extract partners of a set of gene-ids, generate summary statistics over
those neighbours and then cluster the initial gene-ids}
\usage{
run_coex_partner_workflow(gene_ids, gene_statistics, importer,
  gene_universe = NULL, n_partners = 100, ...)
}
\arguments{
\item{gene_ids}{A set of gene identifiers. Should be a subset of
the genes in \code{gene_statistics} and (if defined) \code{gene_universe}.}

\item{gene_statistics}{A table containing the two-tailed p-values,
direction-of-change and gene-ids for all genes that are to be considered
in the analysis (any genes that are not in the \code{gene_universe} are
disregarded).}

\item{importer}{A \code{CoxpresDbAccessor} object.}

\item{gene_universe}{A set of gene identifiers or \code{NULL}. If
defined, each of the gene-identifiers should have data present in a row of
\code{gene_statistics}. If undefined, the universe is taken to be the
intersection of all genes in \code{gene_statistics} and all genes
accessible through the CoxpresDB dataset in \code{importer}. Once defined,
all genes in \code{gene_ids} should be present in \code{gene_universe}.}

\item{n_partners}{The maximum number of partners to pull out for a
given source gene from the CoxpresDb database.}

\item{...}{Other arguments passed to
\code{get_coex_partners}.}
}
\description{
Extract partners of a set of gene-ids, generate summary statistics over
those neighbours and then cluster the initial gene-ids
}
