% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxpresdbr_CoxpresDbPartners.R
\name{.validity_coxpresdb_partners}
\alias{.validity_coxpresdb_partners}
\title{Validity checker for CoxpresDbPartners objects}
\usage{
.validity_coxpresdb_partners(object)
}
\arguments{
\item{object}{A putative CoxpresDbPartners object}
}
\description{
Validity checker for CoxpresDbPartners objects
}
